compintel-tracker/backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
SQLAlchemy==2.0.20
alembic==1.12.0
newspaper3k==0.2.8
werkzeug==3.0.1
psycopg2-binary==2.9.7
ollama
